🎨 Avoid to rerender effects
