+++ src/components/canvas/background/index.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import ParticleField from 'react-particles-webgl'
+
+import { defaultConfig } from '@/helpers/particles/config'
+
+const ParticlesBackground = () => {
+  return (
+    <div id='particle-field' className='relative w-screen h-screen'>
+      <ParticleField config={defaultConfig} />
+    </div>
+  )
+}
+
+export default ParticlesBackground
